data-injector/requirements.txt
# Dependências principais
Flask[async]==3.0.2
pandas==2.1.4
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
asyncio==3.4.3
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTx==1.3.0
aiofiles==23.2.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiohttp==3.9.3
uvicorn==0.27.1
python-multipart==0.0.9

# Dependências de desenvolvimento
pytest==8.0.2
pytest-cov==4.1.0
pytest-asyncio==0.23.5
black==24.1.1
flake8==7.0.0
mypy==1.8.0